Move plants to separate module
