+++ emergence_lib/src/graphics/mod.rs
@@ -2,6 +2,7 @@
 
 use crate::enum_iter::IterableEnum;
 use crate::graphics::terrain::{TerrainSprite, TerrainTilemap};
+use crate::organisms::structures::plants::Plant;
 use bevy::app::{App, CoreStage, Plugin, StartupStage};
 use bevy::asset::AssetServer;
 use bevy::ecs::component::Component;
@@ -26,7 +27,7 @@ use emergence_macros::IterableEnum;
 
 use crate::graphics::produce::{ProduceSprite, ProduceTilemap};
 use crate::graphics::sprites::{IntoSprite, SpriteIndex};
-use crate::organisms::structures::{Fungi, Plant};
+use crate::organisms::structures::Fungi;
 use crate::organisms::units::Ant;
 use crate::simulation::map::MapGeometry;
 use crate::terrain::components::{HighTerrain, PlainTerrain, RockyTerrain};
